+++ packages/espressocash_app/lib/features/outgoing_link_payments/widgets/extensions.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
+import 'package:sentry_flutter/sentry_flutter.dart';
 
 import '../../../di.dart';
 import '../../../ui/loader.dart';
@@ -12,27 +13,29 @@ import '../services/olp_service.dart';
 extension BuildContextExt on BuildContext {
   Future<String> createOLP({required CryptoAmount amount}) =>
       runWithLoader(this, () async {
+        final span = _start(function: 'createOLP');
+
         final payment = await sl<OLPService>().create(
           amount: amount,
           account: sl<MyAccount>().wallet,
         );
+        await span.finish();
 
         return payment.id;
       });
 
   Future<void> cancelOLP({required OutgoingLinkPayment payment}) =>
       runWithLoader<void>(this, () async {
+        final span = _start(function: 'cancelOLP');
         await sl<OLPService>().cancel(
           payment,
           account: sl<MyAccount>().wallet,
         );
-      });
-
-  Future<void> retryOLP({required OutgoingLinkPayment payment}) =>
-      runWithLoader<void>(this, () async {
-        await sl<OLPService>().retry(
-          payment,
-          account: sl<MyAccount>().wallet,
-        );
+        await span.finish();
       });
 }
+
+ISentrySpan _start({required String function}) => Sentry.startTransaction(
+      'OLP Tx Creation',
+      function,
+    );