+++ packages/espressocash_app/lib/features/ramp/services/off_ramp_order_service.dart
@@ -17,13 +17,16 @@ import '../../../data/db/db.dart';
 import '../../../di.dart';
 import '../../accounts/auth_scope.dart';
 import '../../accounts/models/ec_wallet.dart';
+import '../../analytics/analytics_manager.dart';
 import '../../currency/models/amount.dart';
 import '../../currency/models/currency.dart';
 import '../../ramp_partner/models/ramp_partner.dart';
-import '../../tokens/token_list.dart';
+import '../../tokens/data/token_repository.dart';
+import '../../tokens/token.dart';
 import '../../transactions/models/tx_results.dart';
 import '../../transactions/services/resign_tx.dart';
 import '../../transactions/services/tx_sender.dart';
+import '../models/ramp_type.dart';
 import '../models/ramp_watcher.dart';
 import '../partners/coinflow/services/coinflow_off_ramp_order_watcher.dart';
 import '../partners/kado/services/kado_off_ramp_order_watcher.dart';
@@ -40,6 +43,13 @@ typedef OffRampOrder = ({
   FiatAmount? receiveAmount,
   String partnerOrderId,
   Ed25519HDPublicKey? depositAddress,
+  String? moreInfoUrl,
+  String? withdrawAnchorAccount,
+  String? withdrawUrl,
+  String? authToken,
+  String? referenceNumber,
+  CryptoAmount? bridgeAmount,
+  CryptoAmount? refundAmount,
 });
 
 @Singleton(scope: authScope)
@@ -49,17 +59,19 @@ class OffRampOrderService implements Disposable {
     this._client,
     this._sender,
     this._db,
-    this._tokens,
+    this._tokenRepository,
+    this._analytics,
   );
 
   final Map<String, StreamSubscription<void>> _subscriptions = {};
-  final Map<String, RampWatcher> _watchers = {};
+  final Map<String, RampWatcher?> _watchers = {};
 
   final ECWallet _account;
   final EspressoCashClient _client;
   final TxSender _sender;
   final MyDatabase _db;
-  final TokenList _tokens;
+  final TokenRepository _tokenRepository;
+  final AnalyticsManager _analytics;
 
   @PostConstruct(preResolve: true)
   Future<void> init() async {
@@ -74,6 +86,10 @@ class OffRampOrderService implements Disposable {
     final orders = await query.get();
 
     for (final order in orders) {
+      if (order.partner == RampPartner.moneygram) {
+        continue;
+      }
+
       _subscribe(order.id);
       unawaited(_watch(order.id));
     }
@@ -86,6 +102,9 @@ class OffRampOrderService implements Disposable {
       )
       ..where(
         (tbl) => tbl.status.equalsValue(OffRampOrderStatus.cancelled).not(),
+      )
+      ..where(
+        (tbl) => tbl.status.equalsValue(OffRampOrderStatus.refunded).not(),
       );
 
     return query
@@ -98,25 +117,22 @@ class OffRampOrderService implements Disposable {
     final query = _db.select(_db.offRampOrderRows)
       ..where((tbl) => tbl.id.equals(orderId));
 
-    return query.watchSingle().map((row) {
-      final amount = CryptoAmount(
-        value: row.amount,
-        cryptoCurrency: CryptoCurrency(
-          token: _tokens.requireTokenByMint(row.token),
-        ),
-      );
+    return query.watchSingle().asyncMap((row) async {
+      final amount = await _amount(row);
 
-      final fee = row.feeAmount?.let(
-        (amount) {
-          final token = row.feeToken;
+      final fee = await row.feeAmount?.let(
+        (amount) async {
+          final tokenAddress = row.feeToken;
+
+          if (tokenAddress == null) return null;
+
+          final token = await _tokenRepository.getToken(tokenAddress);
 
           if (token == null) return null;
 
           return CryptoAmount(
             value: amount,
-            cryptoCurrency: CryptoCurrency(
-              token: _tokens.requireTokenByMint(token),
-            ),
+            cryptoCurrency: CryptoCurrency(token: token),
           );
         },
       );
@@ -124,14 +140,23 @@ class OffRampOrderService implements Disposable {
       final receiveAmount = row.receiveAmount?.let(
         (it) => Amount(
           value: it,
-          // ignore: avoid-non-null-assertion, checked amount
-          currency: currencyFromString(row.fiatSymbol!),
+          currency: currencyFromString(row.fiatSymbol ?? 'USD'),
         ) as FiatAmount,
+      );
+
+      final bridgeAmount = row.bridgeAmount?.let(
+        (it) => Amount(
+          value: it,
+          currency: Currency.usdc,
+        ) as CryptoAmount,
       );
 
       final depositAddress = row.depositAddress
           .maybeWhere((it) => it.isNotEmpty)
           ?.let(Ed25519HDPublicKey.fromBase58);
+
+      final refundAmount = row.refundAmount
+          ?.let((it) => CryptoAmount(value: it, cryptoCurrency: Currency.usdc));
 
       return (
         id: row.id,
@@ -143,7 +168,14 @@ class OffRampOrderService implements Disposable {
         receiveAmount: receiveAmount,
         partnerOrderId: row.partnerOrderId,
         depositAddress: depositAddress,
+        withdrawAnchorAccount: row.withdrawAnchorAccount,
+        moreInfoUrl: row.moreInfoUrl,
         fee: fee,
+        withdrawUrl: row.withdrawUrl,
+        authToken: row.authToken,
+        referenceNumber: row.referenceNumber,
+        bridgeAmount: bridgeAmount,
+        refundAmount: refundAmount,
       );
     });
   }
@@ -182,9 +214,15 @@ class OffRampOrderService implements Disposable {
         }
       case OffRampOrderStatus.creatingDepositTx:
       case OffRampOrderStatus.depositTxReady:
+      case OffRampOrderStatus.preProcessing:
+      case OffRampOrderStatus.postProcessing:
+      case OffRampOrderStatus.ready:
       case OffRampOrderStatus.sendingDepositTx:
       case OffRampOrderStatus.waitingForPartner:
       case OffRampOrderStatus.failure:
+      case OffRampOrderStatus.processingRefund:
+      case OffRampOrderStatus.waitingForRefundBridge:
+      case OffRampOrderStatus.refunded:
       case OffRampOrderStatus.completed:
       case OffRampOrderStatus.cancelled:
         break;
@@ -203,15 +241,24 @@ class OffRampOrderService implements Disposable {
       case OffRampOrderStatus.depositError:
       case OffRampOrderStatus.insufficientFunds:
         await updateQuery.write(_cancelled);
+      case OffRampOrderStatus.ready:
+        if (order.partner == RampPartner.moneygram) {
+          await updateQuery.write(_processRefund);
+        }
       case OffRampOrderStatus.depositTxRequired:
       case OffRampOrderStatus.creatingDepositTx:
       case OffRampOrderStatus.depositTxReady:
       case OffRampOrderStatus.sendingDepositTx:
       case OffRampOrderStatus.waitingForPartner:
       case OffRampOrderStatus.failure:
+      case OffRampOrderStatus.processingRefund:
+      case OffRampOrderStatus.waitingForRefundBridge:
       case OffRampOrderStatus.completed:
       case OffRampOrderStatus.cancelled:
       case OffRampOrderStatus.depositTxConfirmError:
+      case OffRampOrderStatus.preProcessing:
+      case OffRampOrderStatus.postProcessing:
+      case OffRampOrderStatus.refunded:
         break;
     }
   }
@@ -225,6 +272,7 @@ class OffRampOrderService implements Disposable {
     (SignedTx, BigInt)? transaction,
     FiatAmount? receiveAmount,
     CryptoAmount? fee,
+    required String countryCode,
   }) =>
       tryEitherAsync((_) async {
         {
@@ -253,6 +301,14 @@ class OffRampOrderService implements Disposable {
           _subscribe(order.id);
           await _watch(order.id);
 
+          _analytics.rampInitiated(
+            partner: partner,
+            rampType: RampType.offRamp.name,
+            amount: amount.value.toString(),
+            countryCode: countryCode,
+            id: order.id,
+          );
+
           return order.id;
         }
       });
@@ -264,6 +320,7 @@ class OffRampOrderService implements Disposable {
     required RampPartner partner,
     required BigInt slot,
     FiatAmount? receiveAmount,
+    required String countryCode,
   }) =>
       tryEitherAsync((bind) async {
         {
@@ -276,6 +333,7 @@ class OffRampOrderService implements Disposable {
             depositAddress: '',
             receiveAmount: receiveAmount,
             transaction: (signed, slot),
+            countryCode: countryCode,
           ).letAsync(bind);
         }
       });
@@ -291,6 +349,7 @@ class OffRampOrderService implements Disposable {
       RampPartner.scalex => sl<ScalexOffRampOrderWatcher>(),
       RampPartner.coinflow => sl<CoinflowOffRampOrderWatcher>(),
       RampPartner.rampNetwork ||
+      RampPartner.moneygram || // moneygram orders will not reach this point
       RampPartner.guardarian =>
         throw ArgumentError('Not implemented'),
     }
@@ -306,6 +365,9 @@ class OffRampOrderService implements Disposable {
         case OffRampOrderStatus.depositTxRequired:
         case OffRampOrderStatus.depositError:
         case OffRampOrderStatus.depositTxConfirmError:
+        case OffRampOrderStatus.preProcessing:
+        case OffRampOrderStatus.postProcessing:
+        case OffRampOrderStatus.ready:
         case OffRampOrderStatus.insufficientFunds:
         case OffRampOrderStatus.waitingForPartner:
           return const Stream.empty();
@@ -339,6 +401,9 @@ class OffRampOrderService implements Disposable {
           );
         case OffRampOrderStatus.cancelled:
         case OffRampOrderStatus.failure:
+        case OffRampOrderStatus.processingRefund:
+        case OffRampOrderStatus.waitingForRefundBridge:
+        case OffRampOrderStatus.refunded:
         case OffRampOrderStatus.completed:
           _subscriptions.remove(orderId)?.cancel();
 
@@ -357,25 +422,25 @@ class OffRampOrderService implements Disposable {
   @override
   Future<void> onDispose() async {
     await Future.wait(_subscriptions.values.map((it) => it.cancel()));
-    _watchers.values.map((it) => it.close());
+    _watchers.values.map((it) => it?.close());
     await _db.delete(_db.offRampOrderRows).go();
   }
 
-  CryptoAmount _amount(OffRampOrderRow order) => CryptoAmount(
+  Future<CryptoAmount> _amount(OffRampOrderRow order) async => CryptoAmount(
         value: order.amount,
         cryptoCurrency: CryptoCurrency(
-          token: _tokens.requireTokenByMint(order.token),
+          token: (await _tokenRepository.getToken(order.token)) ?? Token.unk,
         ),
       );
 
   Future<OffRampOrderRowsCompanion> _createTx({
-    required CryptoAmount amount,
+    required Future<CryptoAmount> amount,
     required Ed25519HDPublicKey receiver,
   }) async {
     final dto = CreateDirectPaymentRequestDto(
       senderAccount: _account.address,
       receiverAccount: receiver.toBase58(),
-      amount: amount.value,
+      amount: (await amount).value,
       referenceAccount: null,
       cluster: apiCluster,
     );
@@ -469,5 +534,9 @@ class OffRampOrderService implements Disposable {
 
   static const _depositError = OffRampOrderRowsCompanion(
     status: Value(OffRampOrderStatus.depositTxConfirmError),
+  );
+
+  static const _processRefund = OffRampOrderRowsCompanion(
+    status: Value(OffRampOrderStatus.processingRefund),
   );
 }